requirements.txt
selenium>=4.15.0
pytest>=7.4.0
webdriver-manager>=4.0.1
allure-pytest>=2.13.0
simple-salesforce>=1.12.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
